move to Proposals extra logic of settings validations
